Added bus and add bus functionallity

+++ dotNet5781_2033_0032/dotNet5781_Project_2033_0032/DalApi/DO/Bus.cs
@@ -8,8 +8,8 @@ namespace DO
         public int LicenseNum { get; set; }
         public DateTime FromDate { get; set; }
         public DateTime LastTreatment { get; set; }
-        public float FuelRemaining { get; set; }
-        public float TotalTrip { get; set; }
+        public double FuelRemaining { get; set; }
+        public double TotalTrip { get; set; }
         public Status Status { get; set; }
 
         public const int FullGasTank = 1200;//const the size of full gas tank
